Updated sdk_changelog.txt. Updated building_indigo.txt. Fixed indigo_dll_example/Driver.cpp to work with 4 bytes/pixel alpha channel LDR image. Removed unneeded zlib files from CMakeLists.txt. Added some needed files to sdk_lib/CMakeLists.txt. Made some compile fixes.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@4898 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/AESEncryption.cpp
@@ -107,7 +107,7 @@ std::vector<unsigned char> AESEncryption::decrypt(const std::vector<unsigned cha
 		throw Indigo::Exception("EVP_DecryptFinal_ex failed.");
 
 	int len = plaintext_len + final_len;
-	assert(len <= plaintext.size());
+	assert(len <= (int)plaintext.size());
 	plaintext.resize(len);
 	return plaintext;
 }
